refactor commands and initial advanced columns
