+++ cypress/support/helpers.js
@@ -1,0 +1,63 @@
+/**
+ * Helper function for creating block validation test.
+ *
+ * @param {string} blockName
+ */
+export const blockErrorTest = ( blockName = 'ugb/accordion' ) =>
+	() => {
+		cy.setupWP()
+		cy.newPage()
+		cy.addBlock( blockName )
+		cy.publish()
+		cy.reload()
+	}
+
+/**
+ * Helper function for creating block exist assertion.
+ *
+ * @param {string} blockName
+ * @param {string} selector
+ */
+export const assertBlockExist = ( blockName = 'ugb/accordion', selector = '.ugb-accordion' ) => () => {
+	cy.setupWP()
+	cy.newPage()
+	cy.addBlock( blockName )
+	cy.get( selector ).should( 'exist' )
+}
+
+/**
+ * Helper function for switching designs.
+ *
+ * @param {string} blockName
+ * @param {Array} designs
+ */
+export const switchDesigns = ( blockName = 'ugb/accordion', designs = [] ) => () => {
+	cy.setupWP()
+	cy.newPage()
+	cy.addBlock( blockName )
+	designs.forEach( layout => {
+		cy.openInspector( blockName, 'Layout' )
+		cy.adjustDesign( layout )
+		cy.wait( 1000 )
+		cy.publish()
+		cy.reload()
+	} )
+}
+
+/**
+ * Helper function for switching layouts.
+ *
+ * @param {string} blockName
+ * @param {Array} layouts
+ */
+export const switchLayouts = ( blockName = 'ugb/accordion', layouts = [] ) => () => {
+	cy.setupWP()
+	cy.newPage()
+	cy.addBlock( blockName )
+	layouts.forEach( layout => {
+		cy.openInspector( blockName, 'Layout' )
+		cy.adjustLayout( layout )
+		cy.publish()
+		cy.reload()
+	} )
+}
