+++ cypress/support/index.js
@@ -1,18 +1,3 @@
-// ***********************************************************
-// This example support/index.js is processed and
-// loaded automatically before your test files.
-//
-// This is a great place to put global configuration and
-// behavior that modifies Cypress.
-//
-// You can change the location of this file or turn off
-// automatically serving support files with the
-// 'supportFile' configuration option.
-//
-// You can read more here:
-// https://on.cypress.io/configuration
-// ***********************************************************
-
 Cypress.on( `window:before:load`, win => {
 	cy.stub( win.console, `error`, msg => {
 		// log to Terminal
@@ -23,7 +8,7 @@ Cypress.on( `window:before:load`, win => {
 } )
 
 // Import commands.js using ES2015 syntax:
-import './commands'
+import './commands/index'
 
 // Alternatively you can use CommonJS syntax:
 // require('./commands')
